Rename tester to sim_player
